+++ packages/fast-color-explorer/app/design-system.ts
@@ -6,6 +6,7 @@ import {
 import { createColorPalette } from "@microsoft/fast-components-styles-msft";
 import { ColorRGBA64, parseColorHexRGB } from "@microsoft/fast-colors";
 import { defaultAccentColor, defaultNeutralColor } from "./colors";
+import { StandardLuminance } from "@microsoft/fast-components-styles-msft/dist/utilities/color/neutral-layer";
 
 const neutralPalette: Palette = createColorPalette(parseColorHexRGB(
     defaultNeutralColor
@@ -16,7 +17,7 @@ export const colorsDesignSystem: ColorsDesignSystem = Object.assign(
     {},
     DesignSystemDefaults,
     {
-        backgroundColor: neutralPalette[neutralPalette.length - 1],
+        baseLayerLuminance: StandardLuminance.DarkMode,
         neutralPalette,
         accentPalette: createColorPalette(parseColorHexRGB(
             defaultAccentColor