+++ packages/fast-components-styles-msft/src/design-system/index.ts
@@ -10,6 +10,7 @@ import { withDefaults } from "@microsoft/fast-jss-utilities";
 import { FontWeight } from "../utilities/fonts";
 import designSystemSchema from "./design-system.schema";
 import { accentPalette, neutralPalette } from "../default-palette";
+import { isFunction } from "lodash-es";
 
 export const defaultFontWeights: FontWeight = {
     light: 100,
@@ -161,6 +162,14 @@ export interface DesignSystem {
     neutralFillToggleFocusDelta: number;
 
     /**
+     * The luminance value to base layer recipes on.
+     * Sets the luminance value for the L1 layer recipe in a manner that can adjust to variable contrast.
+     *
+     * Currently defaults to -1 to turn the feature off and use backgroundColor for layer colors instead.
+     */
+    baseLayerLuminance: number; // 0...1
+
+    /**
      * Color swatch deltas for the neutral-fill-card recipe.
      */
     neutralFillCardDelta: number;
@@ -249,6 +258,7 @@ const designSystemDefaults: DesignSystem = {
     neutralFillToggleActiveDelta: -5,
     neutralFillToggleFocusDelta: 0,
 
+    baseLayerLuminance: -1,
     neutralFillCardDelta: 3,
 
     neutralForegroundDarkIndex: 93,
@@ -293,9 +303,7 @@ export function checkDesignSystemResolver<T>(
     arg: T | DesignSystemResolver<T>,
     designSystem: DesignSystem
 ): T {
-    const value: T =
-        typeof arg === "function" ? (arg as DesignSystemResolver<T>)(designSystem) : arg;
-    return value;
+    return isFunction(arg) ? arg(designSystem) : arg;
 }
 
 /**